requirements.txt
certifi==2019.3.9
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4
loggly-python-handler==1.0.0
openpyxl==2.6.2
Pillow==6.0.0
pymarc==3.1.13
PyMySQL==0.9.3
requests==2.21.0
requests-futures==0.9.9
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.24.3
